requirements.txt
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b3
azure-core==1.35.0
azure-identity==1.24.0
azure-monitor-opentelemetry==1.8.0
azure-search-documents==11.5.3
chainlit==2.7.2
openai==1.107.1
opentelemetry-instrumentation-openai==0.47.0
python-dotenv==1.1.1
jsonref==1.1.0
semantic-kernel==1.36.2
agent-framework==1.0.0b251028